refactor: split basic info and images components
